Data Structure and Algorithm
The project focuses on real world data structures and represents them in
a graphically and more interactive format. We have developed a Java
based, easy-to-use, and architecture independent method for creating
animations of data structures run over the Web. By leveraging the
capabilities of Java and the Web, we have built a system that will
provide educators the realistic option of quickly generating algorithm
animations for demonstration use in the classroom or by students
generating quick animations of their programs.