Stack Implementation
A stack is a container of objects that are inserted and removed
according to the last-in first-out (LIFO) principle.
In the pushdown stacks only two operations are allowed: push the item
into the stack, and pop the item out of the stack.
A stack is a limited access data structure - elements can be added and
removed from the stack only at the top.
push adds an item to the top of the stack, pop removes the item from the
top.
A helpful analogy is to think of a stack of books; you can remove only
the top book, also you can add a new book on the top.