Queue Implementation
A queue is a container of objects (a linear collection) that are
inserted and removed according to the first-in first-out (FIFO)
principle.
An excellent example of a queue is a line of students in the food court
of the UC.
New additions to a line made to the back of the queue, while removal (or
serving) happens in the front.
In the queue only two operations are allowed enqueue and dequeue.
Enqueue means to insert an item into the back of the queue, dequeue
means removing the front item.
The picture demonstrates the FIFO access.