Heap Implementation
A heap is a partially sorted binary tree. Although a heap is not
completely in order, it conforms to a sorting principle: every node has
a value less (for the sake of simplicity, we will assume that all
orderings are from least to greatest) than either of its children.
Additionally, a heap is a "complete tree" -- a complete tree is one in
which there are no gaps between leaves. For instance, a tree with a root
node that has only one child must have its child as the left node. More
precisely, a complete tree is one that has every level filled in before
adding a node to the next level, and one that has the nodes in a given
level  filled in from left to right, with no breaks.