Binary Search Tree Implementation
A binary tree is made of nodes, where each node contains a "left"
pointer, a "right" pointer, and a data element. The "root" pointer
points to the topmost node in the tree. The left and right pointers
recursively point to smaller "subtrees" on either side. A null pointer
represents a binary tree with no elements - the empty tree. The formal
recursive definition is: a binary tree is either empty (represented by a
null pointer), or is made of a single node, where the left and right
pointers (recursive definition ahead) each point to a binary tree.